Add future states data collection
Extend StateHeuristic to be able to tarck the evolution of future states so data can be collected and taken into consideration when evaluating the final predicted state. For the winning heuristics is not needed to track any data but the exploratory heuristics needs to track the positions visited to estimation of new positions reached.