Improve ExplorationHeuristic
Update heuristic calculation so 1) new positions visited by the agent are rewarded highly, increasing which each new discovery 2) encourages moving around the map so a position that has visited more times than another penalises more in the heuristic. This tries to make the agent move around the map so there are more chances to discover new positions on its way. 3) gameOver is penalised