Create and initialise MAPElites
Create MAPElites and necessary class to be able to create a map of elites. Included the logic to retrieve the features id from the resulting stats of playing a game as well as the performance of the elite. The current code is mainly the initialisation of the map and it does not include any evolution of parameters yet.
It has been created a test class to make sure the id of the features (the bucket a certain value for a feature) belongs to, which would correspond to the id in the Map elite