Generator and MAPElites
Create new Generator class that contains the methods that generate random values and mutate the weights given to the agents.

MAPElites algorithm has been completed and now the map is initialised by running agents with randomnly assigned weights and then use a stochastic hill climber mutation to get new values for the weights.