Fix weights bug in algorithm
After the map was initialised, running the map elites algorithm did update the weights in the class but the changes were not catched by the heuristic class/agent. The cause was that the reference of the weightList inside the heuristic was removed and the updates were carried out in another reference. The code has been updated to use a unique reference maintained inside MapElites and the weights updates copied the values of the elite taken randomnly, instead of taking a clone (new reference and where the issue came from)