TeamGameplay class
Create new class that contains the internal information about the game, level, file and specifics related to the FW. This class has been created to have the MapElites as transparent to the framework and how the game is run and managed as possible. MapElites makes a call to this class to obtain the gameStats providing a controller and updating the values to the weightlist referenced internally.