Add explorationPercentage stats and percentage features
Include the stats prviding the percentage of the map eplored, for levels with a known size. Including this stat gives more flexibility on features and information as it is possible the user know the size of the map and prefer to create the map of elites with this value instead the number of different positions visited.

Also, the features logic has been updated to take into consideration those stats thata are "percentages" and, therefore, in a range of [0, 1] when calculated. For these values, it is needed to multiply by 100 to be able to use the feature for a map.