MapElitesGameplay configurable using a file
Update logic to make use of a config file (which is hardcoded atm). The main is responsible on setting the variables, provide the valid config file and provide the correspondent files to each method.
A JSON is generated. It contains the config data and the results after running the MAP elites algorithm