Change MAP-elites initialisation criteria
Instead considering a specific number of cells filled in the map, get a number of random configurations to calculate. This chaneg avoids having undesired extremely long initialisations loop and even infinite loops, in the case that the number of cells required is never reach (as this value it is unknown before running the algorithm). The config file has been updated. Also, a new main has been created that simply allows generating a valid config file template