SpritesData and KnowledgeHeuristic
Create new heuristic that will be the base for those behaviours related to sprites: collector, scholar, etc
SpritesData contains all the information related to sprites and interactions and the methods needed to populate these.

Right now the logic allows to discover new sprites, keep track of the collisions (avatar colliding with another sprite) and hits (an sprite created by the avatar collides with a sprite). It is missing including a history of where the collisions and hits happened.