Set SpritesData as static in KnowledgeHeuristic
All heuristics that extends KnowledgeHeuristic should share the same SpritesData as it will be the same for all of them when they are being used in the same game. To avoid having duplicated data and calculations that can slow the agents down, the SpritesData is set as static so it is shared between all of them. It is added a variable that allows a heuristic to update the sprite data, In the case of the MapElites, the TeamBehaviour heuristic will instantiate a KnowledgeHeuristic and be sure the data is available for the rest of the instancies.