Create calculateFieldStats in GameStats
Create new method that allows calcultaing the stats for a specific field instead of having to calculate all of them. This is helpful because in some cases (like the map elited) it wont be needed all stats for all fields and calculating all of them can be time consuming. This method uses field reflecting. IT IS ASSUMED that the name provided is the name of the field without the "Stats" ending. For those fields which stats are calculated directly from an array of data, this array of data is assumed to have the given name, without Stats.