MapElites: update only useful stats during iterations
U[date MapElites logic so only the stats relevant for the algorithm are calculated during its execution. The stats needed are the ones related to the features and performance. When the algorithm ends it is necessary to calculate all the stats for all the data so it is available for printing and serialising.