Add methods to Knowledgeheuristic and SpritesData for Killer and Collector
A series of new methods are necessary to create Killer and Collector heuristics. We need to be able to filter the sprites interacted with and only considered certain stypes. These stypes will be different for Killer and collector. It should also be poosible to get the last time we had an interaction for a certain stype, or series of stypes.