Add temp map elites state data dump
Process and print map elites data every 25 iterations of the algorithm. This would be useful in case the algorithm takes a very long time, to check the status of the map and to have a "screenshot" of how the map elites evolve along time. The data is also saved after initialisation and general info json is also generated before the algorithm starts. These files are generated with a generic name as they are expected to be created on their own folder.