Add exit to AutomatedGameplay.java and create automatedGameplay standalone project
Add exit so the game and executable finishes when the gameplay is over.
Create automatedGameplay standalone project that contains the files necessary
to trigger an automated gameplay of a game of the GVGAI framework. The game
and description of the agent is dependant from the config file provided.
We include to .jar, one where the execution and game finishes when the gameplay
is over, and another one that hangs in the end state and needs to be closed/kill
to finish.