Create AutomatedGameplayExperiments
The new program runs X levels of a game (provided by the config file) a certain number of times to extract their stats.
The config file gives the characteristics the of the agent, game, levels, number of executions for stats calculation, and id and info of the elite used. The games available are Butterflies, Zelda, Digdug and Zelda.